Finished Task 3, Task 4 and Task 6.
In this update, a user is now able to clock on a marker on the map, and a pop-up box appears. This pop-up displays the following information: issued date, work class group, contractor name, community name, and original address. To solve the issue of overlapping markers, the Overlapping Marker Spiderfier plug-n was utilized. Additionally, if a user has already searched for a date range and the markers are shown, if the user searches for a new date range, all existing markers are erased, the map is refreshed and the new matching result is displayed.